Update README instructions and commands list
The README.md was edited to adjust the project setup instructions and update the commands list. The 'testing purposes' phrase was removed from the 'Getting Started' section to provide clarity that the setup is primarily for development. Two new commands, '--list' and '--edit', were added to help users better manage their shortcuts. The 'Running the tests' section was removed due to the absence of automated tests. Instructions for manually editing the configuration file were also added.